refactor: update import paths and rename log and token repositories to align with new module structure
