+++ apps/electron/src/main/modules/mcp-apps-manager/mcp-apps-handler.ts
