+++ apps/electron/src/main/modules/mcp-logger/log-handler.ts
